Moved PyramidQuadrature.
